COMP: Support setting default output_dir and install_destination outside of SlicerExecutionModelConfig.cmake
Currently, it is possible to configure the SlicerExecutionModel so that
default value configure in the SlicerExecutionModelConfig.cmake file are uses.

This is done by configuring the project passing option of the form:
  SlicerExecutionModel_DEFAULT_CLI_*_OUTPUT_DIRECTORY
  SlicerExecutionModel_DEFAULT_CLI_INSTALL_*DESTINATION

This commit extend the idea by allowing to also specify "default" value
independently of so the default value configured in the cache are not
considered. The following variable can be set:

  SlicerExecutionModel_CLI_*_OUTPUT_DIRECTORY
  SlicerExecutionModel_CLI_INSTALL_*DESTINATION