ENH: Add ParameterSerializer support.
Add support for the parameter serialization with
ParameterSerializer (https://github.com/TubeTK/TubeTK-ParameterSerializer).
A module will automatically get new flags to serialize and deserialize
the parameters of the module to/from a JSON file.

Support is enabled by setting SlicerExecutionModel_USE_SERIALIZER to ON,
which will require the specification of the ParameterSerializer build tree,
ParameterSerializer_DIR.