ENH: add testing for the pointfile
Add test for bulk data for the point file, but the parameter file
reading and writing don't check the tags.
Added a simple test case in the module description parser test,
added a sample xml file based on the Slicer Excecution Module
tour xml file to exercise the point and pointfile types.

Issue #2979