COMP: Add missing link dependancy for ITKExpat
Previous transitive contamination of the library
linkage would cause nearly all libraries to depend
on each other.  As we fix this in ITK proper (work
by Zach Williamson), we need to set the COMPONENTS
to include the items that we use.

In this example, we need ITKExpat library functions,
so including ITKIOExpat no longer transatively includes
ITKExpat (third party) publicly to child projects.  This
means that it must be included directly. NOTE, Had we used
only ITKIOExpat functions, this would not be necessary.