ENH: Reorganize docker files to support testing against multiple ITK version
This commit extends Docker support allowing to have a Dockerfile for
each version of ITK.

Each Dockerfile is expected to be in a directory of this form:

  test/Docker-<IDENTIFIER>

containing files "Dockerfile", "build.sh" and "run.sh".

<IDENTIFIER> is a free form string allowing to loosely describe the
the properties of the associated ITK build tree. This identifier will be
used to specify the docker image tag in order to differenciate images
regarding to their ITK build tree version. (It corresponds to
variable 'image_tag' in build.sh)

test/Docker-<IDENTIFIER>/build.sh: The ITK build tree is contained
in the docker container that can be generated using the "build.sh"
script and the associated "Dockerfile".

test/Docker/test.sh: This script will

  (a) build SlicerExecutionModel against the ITK build tree available in
      the container obtained using "build.sh"

  (b) run SlicerExecutionModel tests

  Note that results of step (a) and (b) are both submitted to CDash.

test/Docker-<IDENTIFIER>/run.sh: Allow to configure, build and test
SlicerExecutionModel inside the docker container.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>