ENH: Add Relevant BUILDNAME for CDash submission
This commit enables continuous integration testing with docker to
submit results on the dashboard with a more relevant build name.
This build name corresponds to the docker image tag that was previously
set up in PR #49 to specify the build option used by the image.
The image tag is extracted from the name of the Docker directory that
contains the Dockerfile. This directory must be named according to this
convention : Docker-$DOCKER_TAG.
To provide a relevant build name to the image, $DOCKER_TAG has to be set
as an environment variable in the circle.yml file where it is passed as
an argument when running the test.sh script.