ENH: cxx and executable can have different names.
Previously, each CLI needed to have a cxx file called the same as the given executable name. This was restricting the possibility to name an executable differently based, for example, on configuration or compilation options. This patch introduced a new option to disable this requirement. This allows SEMMacroBuildCLI to stay backward compatible while allowing more flexibiliy in the naming of the cxx files and the executable [1].

[1] https://docs.google.com/a/kitware.com/document/d/1Wkl7RFVb87ZbcgZnSQ4nmyM2bFEQUN1xT-Q5ALLJh8w/edit?usp=sharing