STYLE: ModuleDescription: Remove unused Alternative Type/Target/Location
Originally introduced in Slicer r4550 (ENH: Added AlternativeType,
AlternativeTarget, AlternativeLocation so that two different versions
of a module can be made available. [...]), it has never been used in the
CLI logic.

Here is some background information reported by Jim Miller <millerjv@ge.com>,
the original author of the Alternative Type/Target/Location feature. This
could be useful for people who would like to reintroduce it.

The Alternative fields were introduced so that applications could manage both a
shared object version and an executable version of a CLI. The vision was that
the shared object version would be used by the app in interactive sessions. Then,
the executable version would be used when spawning a batch process from the app.

We did use these fields in a prototype for a "Personal Slicer Cloud". Instead of
executing a CLI locally, the user had the option of sending the request to the
cloud, where the CLI would be run, and the results downloaded back into the
Slicer session.

Reviewed-by: Jim Miller <millerjv@ge.com>