ENH: ModuleDescription: Add support for lazily setting the target
To support lazy loading of shared library modules, the library loading
and resolution of symbols should be delayed until it is needed.

This commit updates ModuleDescription so that a callback in charge of
loading the library and resolving symbols can optionally be set. It will
be invoked if no target has already been set when "GetTarget()" is called.