ENH: Improve JsonCpp support for SlicerExecutionModel
There are a few items here to improve the JsonCpp support for
SlicerExecutionModel:

- FindJsonCpp.cmake:
Add /src/lib_json directory to the list of directories searches for the
json library. This is the default build directory for the jsoncpp library.

- GenerateCLP/CMakeLists.txt:
Add missing include_directory and missing library linkage to JsonCpp.

- SlicerExecutionModelConfig.cmake.in:
Add missing checks for JsonCpp. Also add definition for windows compilation