ENH: Improve support for empty arrays
The parsing was including the file from the JSON file even if the array
was empty. To remediate to this, we test for the array size to make
sure it's not empty before adding the flag and the following values.

Testing is improved as well by adding an argument with the "multiple" flag
and making sure it's serialized/deserialized correctly (it should always
be empty).