ENH: ParameterSerializer: Fix multiple parameters handling for deserialization
The code wasn't handling multiple parameters correctly and crashing when
trying to deserialize arrays of arrays. This commit addresses that.

Arguments using "multiple=true" can either be flagged or positional.
 - When they are positional, they must be the last argument of the CLI. In
this case, the infrastructure already handles this.
 - When they are flagged, we need to detect that the argument allows for
multiple flag/value pairs. To do so, we introduce the map called
deserializedMultipleArgsMap. It stores a map of flag to vector of values.
Json can populate that map based on the values it reads. Otherwise the
command line parsing will populate this map.
When an argument using "multiple=true" is present in the JSON and in the
command line, simirarly than with other arguments, the values of the JSON are
ignored.

Testing for all the types that support the multiple argument is added.