ENH: GenerateCLP: Force GenerateCLPLauncher rebuild on GenerateCLP change
Previously, the launcher did not rebuild when the GenerateCLP executable
changed. This caused other application that have a dependency on the
launcher to NOT regenerate their CLP header files. The user would have
to specifically ask for a rebuild.

By forcing the launcher to rebuild whenever GenerateCLP has changed, we
make sure that the applications will regenerate their CLP header file.