ENH: Add forward reference support
It was possible to define references before, but they have special meaning (set parent transform, etc), and they are defined "in reverse", i.e. in the referenced object.
This change adds support for generic, forward references, which allow specifying any type of reference between two nodes. For example in a registration module, the output transform can specify node references to the moving and fixed images.

An example element (under a parameter element) is:
  <reference role="spatialRegistrationMoving" parameter="movingVolume"/>