ENH: Allow applications use UTF-8 code page
SlicerExecutionModel_USE_UTF8 is OFF (this is the default, for backward compatibility) then standard Windows API calls expects strings to be stored in the ANSI code page configured on the system.

If SlicerExecutionModel_USE_UTF8 is ON then all std::string or char buffers are expected to contain UTF-8 encoded string. This is what was the default on Mac OSX and Linux and starting from Windows Version 1903 (May 2019 Update) it has been made available as an option on Windows as well. This is where VTK and ITK (and most other software libraries) are moving towards, too.